Make algorithm generic between MultiPolygon and Polygon
