+++ geom/type_multi_polygon.go
@@ -430,6 +430,5 @@ func (m MultiPolygon) Force2D() MultiPolygon {
 
 // PointOnSurface returns a Point on the interior of the MultiPolygon.
 func (m MultiPolygon) PointOnSurface() Point {
-	// TODO
-	return Point{}
+	return pointOnAreaSurface(m)
 }
